+++ CHANGELOG.md
@@ -21,3 +21,8 @@
 * Added Buff onDraw callback
 * Buff.get_stack_count : Now returns "0" if it returned "nil" previously
 * Item.spawn_drop : Now returns the dropped instance
+* Added Object module
+    * Object.find
+    * Object.create
+    * Object.spawn
+    * Object.add_callback